Re-organized chat files to their homes
