+++ lib/chat/StreamChatClient.js
